Feat:nuevas caracteristicas

+++ src/main/java/com/mycompany/tallerexpress/dao/OrdenesServicioDao.java
@@ -1,14 +1,24 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
 package com.mycompany.tallerexpress.dao;
 
 import com.mycompany.tallerexpress.model.OrdenesServicio;
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface OrdenesServicioDao {
-    OrdenesServicio registrarOrdenTransaccional(OrdenesServicio orden, Integer repuestoId, Integer cantidad) throws Exception;
-    void actualizarEstadoYCostoTransaccional(Integer ordenId, String nuevoEstado, Double costoTotal) throws Exception;
-    List<OrdenesServicio> findByVehiculo(Integer vehiculoId) throws Exception;
+
+    OrdenesServicio registrarOrdenTransaccional(
+            OrdenesServicio orden,
+            Integer repuestoId,
+            Integer cantidad
+    ) throws Exception;
+
+    void actualizarEstadoYCostoTransaccional(
+            Integer ordenId,
+            String nuevoEstado,
+            BigDecimal costoTotal
+    ) throws Exception;
+
+    List<OrdenesServicio> findByVehiculo(
+            Integer vehiculoId
+    ) throws Exception;
 }
